+++ CHANGELOG.md
@@ -6,4 +6,16 @@
 - Designed a configManager component
 - Designed a fileSee component
 - Designed a passwordStore component
+
+### 2024.10.07
+### Added
+
+- Created constructor methods
+- Implemented FileSee component
+    - displayNumberedFileTree
+    - listFiles
+    - currentDir
+    - More to come: (select dir/change dir)
+- Created main method to demonstrate the viability of the methods, working on
+Fedora Linux 40 Kernel: 6.11.0-400.asahi.fc40.aarch64+16k, 64bit
 ```